Final Build Submission
Added introduction text with a message dialogue box to prompt the user as to whether they want to resume the console applicaiton. Removed switch statement for Readkey method as instead utilised an easier method with ConsoleKey response in a do loop. New jokes are able to be read on screen. Made an attempt at caching.